#61: Move ToModel converters to different location
