+++ Api/ShoppingList.Api.Infrastructure/Converters/ToDomain/ShoppingListConverter.cs
@@ -12,7 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ProjectHermes.ShoppingList.Api.Infrastructure.Converters
+namespace ProjectHermes.ShoppingList.Api.Infrastructure.Converters.ToDomain
 {
     public class ShoppingListConverter : IToDomainConverter<Entities.ShoppingList, IShoppingList>
     {
